📝 chore: move RatingStar component to common directory
